requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
